Move components into their own directory, updating their respective imports and invocations
