+++ src/App.jsx
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom'
-import RegisterScreen from './RegisterScreen'
-// import HomeScreen from './HomeScreen'
-import Dashboard from './Dashboard'
-import Transacciones from './Transacciones'
-import TareasHabitos from './TareasHabitos'
+import RegisterScreen from './components/RegisterScreen'
+// import HomeScreen from './components/HomeScreen'
+import Dashboard from './components/Dashboard'
+import Transacciones from './components/Transacciones'
+import TareasHabitos from './components/TareasHabitos'
 import logo from './img/color-bw-03.svg'
 // import { BarChart, Bar, XAxis, YAxis, Tooltip, CartesianGrid } from 'recharts'
 import './App.css'
