+++ src/components/ResumenGeneral.jsx
